+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/plan/util/OverAggregateUtil.scala
@@ -89,7 +89,7 @@ object OverAggregateUtil {
       }
       val fields = new JArrayList[RelFieldCollation]()
       for (field <- groupCollation ++ orderCollation) {
-        fields.add(RelFieldCollationUtil.of(field._1, field._2, field._3))
+        fields.add(FlinkRelOptUtil.ofRelFieldCollation(field._1, field._2, field._3))
       }
       RelCollations.of(fields)
     } else {
